+++ core/src/main/kotlin/com/github/ludmilaraskina/config/AmqpLoggingConfig.kt
@@ -3,7 +3,6 @@ package com.github.ludmilaraskina.config
 import com.github.ludmilaraskina.MessageLoggingProcessor
 import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory
 import org.springframework.amqp.rabbit.core.RabbitTemplate
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration
@@ -12,7 +11,6 @@ import javax.annotation.PostConstruct
 @Configuration
 @ConditionalOnProperty("spring.amqp.log.enabled")
 @EnableConfigurationProperties(AmqpLoggingProperties::class)
-@ConditionalOnClass(RabbitTemplate::class)
 class AmqpLoggingConfig(private val rabbitTemplate: RabbitTemplate,
                         private val rabbitListenerContainerFactory: SimpleRabbitListenerContainerFactory) {
