+++ docker-compose.yaml
@@ -1,0 +1,7 @@
+version: "3"
+services:
+  rabbitmq:
+    image: rabbitmq:management
+    ports:
+      - "5674:5672"
+      - "15674:15672"
