+++ example/receiver/build.gradle.kts
@@ -1,14 +1,14 @@
-apply(plugin = "org.jetbrains.kotlin.plugin.spring")
 apply(plugin = "org.springframework.boot")
-apply(plugin = "io.spring.dependency-management")
 
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-amqp")
+    implementation("com.github.ludmilaraskina:spring-amqp-logger:$version")
+    implementation(project(":example:api"))
+
     testImplementation("org.springframework.boot:spring-boot-starter-test") {
         exclude("org.junit.vintage", "junit-vintage-engine")
     }
     testImplementation("org.springframework.amqp:spring-rabbit-test")
-    implementation(kotlin("stdlib-jdk8"))
 }
 
 tasks.test {
