+++ example/receiver/src/main/kotlin/com/github/ludmilaraskina/receiver/Receiver.kt
@@ -1,0 +1,19 @@
+package com.github.ludmilaraskina.receiver
+
+import com.github.ludmilaraskina.api.ExampleMessage
+import com.github.ludmilaraskina.receiver.config.queueName
+import org.springframework.amqp.rabbit.annotation.RabbitListener
+import org.springframework.stereotype.Component
+import java.util.concurrent.CountDownLatch
+
+
+@Component
+class Receiver {
+    val latch = CountDownLatch(1)
+
+    @RabbitListener(queues = [queueName])
+    fun receiveMessage(message: ExampleMessage) {
+        println("Received <$message>")
+        latch.countDown()
+    }
+}
