+++ example/receiver/src/main/kotlin/com/github/ludmilaraskina/receiver/config/AmqpConfig.kt
@@ -1,0 +1,38 @@
+package com.github.ludmilaraskina.receiver.config
+
+import org.springframework.amqp.core.Binding
+import org.springframework.amqp.core.BindingBuilder
+import org.springframework.amqp.core.Queue
+import org.springframework.amqp.core.TopicExchange
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter
+import org.springframework.amqp.support.converter.MessageConverter
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+
+const val topicExchangeName = "spring-boot-exchange"
+const val queueName = "spring-boot"
+
+@Configuration
+class AmqpConfig {
+
+    @Bean
+    fun queue(): Queue? {
+        return Queue(queueName, false)
+    }
+
+    @Bean
+    fun exchange(): TopicExchange? {
+        return TopicExchange(topicExchangeName)
+    }
+
+    @Bean
+    fun binding(queue: Queue?, exchange: TopicExchange?): Binding? {
+        return BindingBuilder.bind(queue).to(exchange).with("foo.bar.#")
+    }
+
+    @Bean
+    fun messageConverter(): MessageConverter {
+        return Jackson2JsonMessageConverter()
+    }
+}
