--- example/receiver/src/main/kotlin/com/github/ludmilaraskina/sender/config/AmqpConfig.kt
@@ -1,49 +1,0 @@
-package com.github.ludmilaraskina.sender.config
-
-import com.github.ludmilaraskina.sender.receiver.Receiver
-import org.springframework.amqp.core.Binding
-import org.springframework.amqp.core.BindingBuilder
-import org.springframework.amqp.core.Queue
-import org.springframework.amqp.core.TopicExchange
-import org.springframework.amqp.rabbit.connection.ConnectionFactory
-import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer
-import org.springframework.amqp.rabbit.listener.adapter.MessageListenerAdapter
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-const val topicExchangeName = "spring-boot-exchange"
-const val queueName = "spring-boot"
-
-@Configuration
-class AmqpConfig {
-
-    @Bean
-    fun queue(): Queue {
-        return Queue(queueName, false)
-    }
-
-    @Bean
-    fun exchange(): TopicExchange {
-        return TopicExchange(topicExchangeName)
-    }
-
-    @Bean
-    fun binding(queue: Queue?, exchange: TopicExchange?): Binding {
-        return BindingBuilder.bind(queue).to(exchange).with("foo.bar.#")
-    }
-
-    @Bean
-    fun container(connectionFactory: ConnectionFactory, listenerAdapter: MessageListenerAdapter): SimpleMessageListenerContainer {
-        val container = SimpleMessageListenerContainer()
-        container.connectionFactory = connectionFactory
-        container.setQueueNames(queueName)
-        container.setMessageListener(listenerAdapter)
-        return container
-    }
-
-    @Bean
-    fun listenerAdapter(receiver: Receiver?): MessageListenerAdapter? {
-        return MessageListenerAdapter(receiver, "receiveMessage")
-    }
-
-}
