--- example/receiver/src/main/kotlin/com/github/ludmilaraskina/sender/receiver/Receiver.kt
@@ -1,16 +1,0 @@
-package com.github.ludmilaraskina.sender.receiver
-
-import org.springframework.stereotype.Component
-import java.util.concurrent.CountDownLatch
-
-
-@Component
-class Receiver {
-    val latch = CountDownLatch(1)
-
-    fun receiveMessage(message: String) {
-        println("Received <$message>")
-        latch.countDown()
-    }
-
-}
