+++ example/sender/build.gradle.kts
@@ -3,12 +3,12 @@ apply(plugin = "org.springframework.boot")
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-amqp")
     implementation("com.github.ludmilaraskina:spring-amqp-logger:$version")
+    implementation(project(":example:api"))
 
     testImplementation("org.springframework.boot:spring-boot-starter-test") {
         exclude("org.junit.vintage", "junit-vintage-engine")
     }
     testImplementation("org.springframework.amqp:spring-rabbit-test")
-    implementation(kotlin("stdlib-jdk8"))
 }
 
 tasks.test {
