+++ settings.gradle.kts
@@ -3,3 +3,4 @@ rootProject.name = "spring-amqp-logger"
 include("core")
 include("example:sender")
 include("example:receiver")
+include("example:api")
